requirements.txt
# Core ML/DL Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Experiment Tracking & Visualization
wandb>=0.15.0
matplotlib>=3.5.0
seaborn>=0.11.0

# UI & Demo
gradio>=4.0.0

# Development & Code Quality
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Optional: For advanced features
tensorboard>=2.10.0  # Alternative to W&B
pillow>=9.0.0       # Image processing
tqdm>=4.64.0        # Progress bars